requirements.txt
google-generativeai==0.3.1
python-dotenv==1.0.0
pyzotero==1.5.18
beautifulsoup4
openai
langchain-google-genai